requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
streamlit
google-generativeai
watchdog
python-dotenv
requests
